requirements.txt
# Core ML
torch>=2.2.0
torch-geometric>=2.5.0

# MARL Environment
pettingzoo>=1.24.0
gymnasium>=0.29.0
supersuit>=3.9.0

# Conformal Prediction
mapie>=0.9.0
scikit-learn>=1.4.0

# Knowledge Graph
neo4j>=5.0.0
sentence-transformers>=2.5.0

# Google Cloud / Gemini
google-genai>=1.30.0
google-cloud-aiplatform>=1.40.0
firebase-admin>=6.4.0

# Utilities
tensorboard>=2.16.0
numpy>=1.26.0
pyyaml>=6.0.0
python-dotenv>=1.0.0
requests>=2.31.0
flask>=3.0.0
flask-cors>=4.0.0

# Dev/Test
pytest>=8.0.0
